# FamilyBenefits-api-openapi.yaml
openapi: 3.0.3
info:
  title: Family Benefits
  description: rest-api для приложения "Пособия для семьи с детьми"
  version: 1.0.0
  contact:
    email: smegovic@gmail.com
servers:
  - url: https://localhost
...
paths:
...
  /benefit:
    post:
      tags:
        - benefit
      summary: Добавляет новое пособие
      description:
        Добавляет новое пособие.
        Для администратора.
        Необходима авторизация.
      operationId: addBenefit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BenefitAdd'
        required: true
      responses:
        '201':
          description: Пособие добавлено
          content: { }
        '400':
          description: Пособие с указанными данными уже существует
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    put:
      tags:
        - benefit
      summary: Обновляет пособие
      description:
        Обновляет пособие.
        Для администратора.
        Необходима авторизация.
      operationId: updateBenefit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BenefitUpdate'
        required: true
      responses:
        '201':
          description: Пособие обновлено
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Пособие с указанными данными не найдено
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /benefit/{id}:
    delete:
      tags:
        - benefit
      summary: Удаляет пособие
      description:
        Удаляет пособие.
        Для администратора.
        Необходима авторизация.
      operationId: deleteBenefit
      parameters:
        - name: id
          in: path
          description: ID пособия
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '201':
          description: Пособие удалено
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Пособие с указанными данными не найдено
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    get:
      tags:
        - benefit
      summary: Возвращает информацию о пособии
      description:
        Возвращает информацию о пособии.
      operationId: getBenefit
      parameters:
        - name: id
          in: path
          description: ID пособия
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '200':
          description: Данные о пособии получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenefitInfo'
        '404':
          description: Пособие с указанным ID не найдено
          content: { }
      security: [ ]
...
  /benefit/allpartial:
    get:
      tags:
        - benefit
      summary: Возвращает множество пособий, в которых нет городов, учреждений или критерий
      description:
        Возвращает множество пособий, в которых нет городов, учреждений или критерий.
        Для администратора.
        Необходима авторизация.
      operationId: getPartialBenefits
      responses:
        '200':
          description: Множество пособий получено
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BenefitInfo'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /benefit/initdata:
    get:
      tags:
        - benefit
      summary: Возвращает дополнительные данные для пособия
      description:
        Возвращает дополнительные данные для добавления или обновления пособия.
        Данные содержат в себе множества кратких информаций о городах, полных критериях и учреждениях
        Для администратора.
        Необходима авторизация.
      operationId: getBenefitInitData
      responses:
        '200':
          description: Дополнительные данные пособия получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenefitInitData'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
...
  /criterion/{id}:
    delete:
      tags:
        - criterion
      summary: Удаляет критерий
      description:
        Удаляет критерий.
        Для администратора.
        Необходима авторизация.
      operationId: deleteCriterion
      parameters:
        - name: id
          in: path
          description: ID критерия
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '201':
          description: Критерий удален
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Критерий с указанными данными не найден
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    get:
      tags:
        - criterion
      summary: Возвращает информацию о критерии
      description:
        Возвращает информацию о критерии.
      operationId: getCriterion
      parameters:
        - name: id
          in: path
          description: ID критерия
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '200':
          description: Данные о критерии получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriterionInfo'
        '404':
          description: Критерий с указанным ID не найден
          content: { }
      security: [ ]
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CriterionInfo:
      description: Информация о критерии для подбора пособий
      properties:
        id:
          description: ID критерия
          type: string
          example: td56387fhj
        name:
          description: Название критерия
          type: string
          example: name
        info:
          description: Информация критерия
          type: string
          example: info
        shortCriterionType:
          $ref: '#/components/schemas/ObjectShortInfo'
        shortBenefitSet:
          description: Множество кратких информаций о пособиях критерия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
...
    BenefitInfo:
      description: Информация о пособии
      properties:
        id:
          description: ID пособия
          type: string
          example: td56387fhj
        name:
          description: Название пособия
          type: string
          example: name
        info:
          description: Информация пособия
          type: string
          example: info
        documents:
          description: Документы для получения пособия
          type: string
          example: documents
        shortCitySet:
          description: Множество кратких информаций о городах пособия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
        shortInstitutionSet:
          description: Множество кратких информаций об учреждениях пособия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
        criterionSet:
          description: Множество информаций о критериях пособия
          type: array
          items:
            $ref: '#/components/schemas/CriterionInfo'
    BenefitAdd:
      description: Объект запроса для добавления пособия
      properties:
        name:
          description: Название пособия
          type: string
          example: name
        info:
          description: Информация пособия
          type: string
          example: info
        documents:
          description: Документы для получения пособия
          type: string
          example: documents
        idCitySet:
          description: Множество ID городов пособия
          type: array
          items:
            description: ID города
            type: string
            example: td56387fhj
        idInstitutionSet:
          description: Множество ID учреждений пособия
          type: array
          items:
            description: ID учреждения
            type: string
            example: td56387fhj
        idCriterionSet:
          description: Множество ID критерий пособия
          type: array
          items:
            description: ID критерия
            type: string
            example: td56387fhj
    BenefitUpdate:
      description: Объект запроса для обновления пособия
      properties:
        id:
          description: ID пособия
          type: string
          example: td56387fhj
        name:
          description: Название пособия
          type: string
          example: name
        info:
          description: Информация пособия
          type: string
          example: info
        documents:
          description: Документы для получения пособия
          type: string
          example: documents
        idCitySet:
          description: Множество ID городов пособия
          type: array
          items:
            description: ID города
            type: string
            example: td56387fhj
        idInstitutionSet:
          description: Множество ID учреждений пособия
          type: array
          items:
            description: ID учреждения
            type: string
            example: td56387fhj
        idCriterionSet:
          description: Множество ID критерий пособия
          type: array
          items:
            description: ID критерия
            type: string
            example: td56387fhj
    BenefitInitData:
      description: Дополнительные данные для добавления или обновления пособия
        Содержат в себе множества кратких информаций о городах, учреждениях и полных критериях
      properties:
        shortCitySet:
          description: Множество кратких информаций о городах
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
        shortInstitutionSet:
          description: Множество кратких информаций об учреждениях
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
        criterionSet:
          description: Множество информаций о критериях
          type: array
          items:
            $ref: '#/components/schemas/CriterionInfo'
    ObjectShortInfo:
      description: Краткая информация об объекте
      properties:
        idObject:
          description: ID объекта
          type: string
          example: td56387fhj
        nameObject:
          description: Название объекта
          type: string
          example: name